refactor: test step 패키지 변경
