+++ src/test/java/nextstep/subway/acceptance/LineAcceptanceTest.java
@@ -3,6 +3,7 @@ package nextstep.subway.acceptance;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import nextstep.subway.acceptance.dto.LineTestRequest;
+import nextstep.subway.acceptance.step.LineTestStep;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
