+++ src/test/java/nextstep/subway/acceptance/StationAcceptanceTest.java
@@ -2,6 +2,7 @@ package nextstep.subway.acceptance;
 
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
+import nextstep.subway.acceptance.step.StationTestStep;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
